Update homepage
Removes icons, as they are not rendering correctly.
Styles the introduction header and adds new paragraph with link to About
Us page, in the introduction section for easier accessibility.
Removes Contact Us and About Us from Links section as they can be found
in the footer.
Makes introduction message easier to read.